requirements.txt
Django==1.8.3
django-admin-bootstrapped==2.5.4
django-bootstrap3==6.2.2
Pillow==2.9.0
psycopg2==2.6.1
sorl-thumbnail==12.3
